Update README.md
iReady is awful. It's the worst education tool anyone could ever use. I'm fed up with being forced to mindlessly watch the result of a greedy corporation that doesn't try in any way to make their product enjoyable, or even acceptable, for an hour and a half every single week. This repository is a collection of hacks that ensure nobody has to suffer through iReady ever again. The current version has a lesson and quiz skipper, minutes hack, and an experimental diagnostic hack. There is also a lesson viewer that can show you the lesson data and files of many iReady lessons. You can view older versions of hacks here. There is also a Chrome extension that makes using these hacks a lot easier. It is an unpackaged extension, meaning you can install it like this. 111